Refactor OpenvisualizerApp
