+++ bin/logging.conf
@@ -167,7 +167,7 @@ qualname=ParserPrintf
 
 [logger_ParserIEC]
 level=INFO
-handlers=std,console
+handlers=std, console
 propagate=0
 qualname=ParserIEC
 
@@ -191,7 +191,7 @@ qualname=OpenLbr
 
 [logger_RPL]
 level=INFO
-handlers=std
+handlers=std, console
 propagate=0
 qualname=RPL
 
@@ -213,17 +213,11 @@ handlers=std, console
 propagate=0
 qualname=openVisualizerApp
 
-[logger_openVisualizerGui]
+[logger_OpenTui]
 level=INFO
-handlers=std
+handlers=std, console
 propagate=0
-qualname=openVisualizerGui
-
-[logger_openVisualizerCli]
-level=INFO
-handlers=std
-propagate=0
-qualname=openVisualizerCli
+qualname=OpenTui
 
 [logger_openVisualizerWeb]
 level=INFO
